requirements.txt
fastapi[standard]
uvicorn
python-multipart
PyPDF2
pydantic
pydantic-settings
openai
python-dotenv
httpx
youtube-transcript-api
numpy
scikit-learn
yt-dlp


# Core dependencies
PyPDF2>=3.0.0
torch>=2.0.0
transformers>=4.46.0
accelerate>=0.27.0
rich>=13.0.0
ipywidgets>=8.0.0
tqdm>=4.66.0

# Optional but recommended
jupyter>=1.0.0
ipykernel>=6.0.0

# Warning handling
warnings>=0.1.0

sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart



sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental